Added two design proposals to semantic_logger.go
