--- newlogger.go
@@ -1,26 +1,0 @@
-package skynet
-
-import (
-	"time"
-)
-
-type Exception struct {
-	Exception string `json:"exception"`
-	Message   string `json:"message"`
-	Backtrace string `json:"backtrace"`
-}
-
-type Payload struct {
-	Name        string        `json:"name"`
-	Application string        `json:"application"`
-	HostName    string        `json:"host_name"`
-	ThreadName  string        `json:"thread_name"`
-	Message     string        `json:"message"`
-	Tags        []string      `json:"tags"`
-	PID         int64         `json:"pid"`
-	Level       string        `json:"level"`
-	Time        time.Time     `json:"time"`
-	Duration    time.Duration `json:"duration"`
-	Table       string        `json:"table"`
-	Action      string        `json:"action"`
-}
